requirements.txt
yfinance
pyyaml